gsg-api/requirements.txt
asgiref==3.5.2
autopep8==2.0.0
certifi==2022.6.15
charset-normalizer==2.1.0
distlib==0.3.6
Django==4.0.6
django-cors-headers==3.13.0
django-filter==22.1
django-unixtimestampfield==0.5.0
djangorestframework==3.14.0
ExifRead==3.0.0
filelock==3.8.0
gunicorn==20.1.0
idna==3.3
Markdown==3.4.1
numpy==1.23.5
pathlib2==2.3.7.post1
Pillow==9.2.0
platformdirs==2.5.4
pycodestyle==2.9.1
python-dotenv==0.20.0
pytz==2022.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
urllib3==1.26.11
virtualenv==20.17.0
